Refactor meal planning UI, implement dedicated recipe selection screen, and update data models for production API compatibility

+++ app/src/main/java/com/teamconfused/planmyplate/domain/model/Grocery.kt
@@ -17,6 +17,6 @@ data class GroceryList(
 data class GroceryListItem(
     val id: Int? = null,
     val ingredient: Ingredient? = null,
-    val quantity: Int? = null,
+    val quantity: Double? = null,
     val unit: String? = null
 )
